+++ packages/modules/auth/src/resolvers/mutations/revoke-customer-token-resolver.ts
@@ -5,7 +5,10 @@ import { getDecodedAuthToken, getVerifiedRefreshToken } from '../../utils';
 
 export const revokeCustomerTokenResolver = {
     resolve: async (_root, args, context, _info) => {
-        const authToken = context.headers.authorization;
+        const authToken = context.request.headers.get('authorization');
+        if (!authToken) {
+            throw new GraphqlError('Missing authorization header', 'authorization');
+        }
 
         const { refresh_token } = args;
 
@@ -19,7 +22,7 @@ export const revokeCustomerTokenResolver = {
             throw new GraphqlError('There was an issue decoding the access token', 'authorization');
         }
 
-        const { bc_customer_id } = decodedAuthToken;
+        const { customer_id } = decodedAuthToken;
 
         /* The "getVerifiedRefreshToken" recreates the old refresh token. If this can't be done
          * then the refresh token may not belong to the user. */
@@ -34,7 +37,7 @@ export const revokeCustomerTokenResolver = {
 
         const authService: AuthService = context.injector.get(AuthService);
 
-        const response = await authService.removeUserAuth(bc_customer_id, refresh_token);
+        const response = await authService.removeUserAuth(customer_id, refresh_token);
 
         if (response instanceof Error) {
             throw new GraphqlError(