+++ packages/modules/auth/src/types/index.ts
@@ -1,7 +1,19 @@
 import { BatchWriteItemCommandOutput } from '@aws-sdk/client-dynamodb/dist-types/commands/BatchWriteItemCommand';
 
+// Extend Global context type see: https://the-guild.dev/graphql/modules/docs/essentials/type-safety#shaping-context-type
+declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
+    namespace GraphQLModules {
+        interface GlobalContext {
+            request: Request;
+        }
+    }
+}
+
+export type CustomerId = number | string;
+
 export type decodedAccessToken = {
-    bc_customer_id: number;
+    customer_id: CustomerId;
     exp: number;
     refresh_expiry: number;
 };