+++ packages/modules/auth/release-notes.md
@@ -1,6 +1,20 @@
 # Auth Module Release Notes
 
-## auth-module-1.0.4
+## auth-module 1.0.0
+
+### Overview of changes
+
+Removes BigCommerce specific login logic from `generate-customer-token` and instead defines
+a stub LoginService class. This service must be extended by other modules and provided via dependency injection.
+
+This module can then be agnostic of the login implementation required for a specific client/platofmr and handle only the logic required for generating, refreshing, caching, and revoking tokens.
+
+#### Tickets
+
+- MI-73: Refactor Authentication GraphQL Module so it isn't tied to the BigCommerce Module
+  - https://aligent.atlassian.net/browse/MI-73
+
+## auth-module-0.0.4
 
 ### Overview of changes
 
@@ -16,7 +30,3 @@ a refresh token has fully expired.
   - https://aligent.atlassian.net/browse/MI-17
 - AM-1554: [QA] The user's login session is getting expired (after 15 minutes) during the time when user is still accessing the website
   - https://matsuyaginza.atlassian.net/browse/AM-1554
-
-#### Pull requests
-
-- https://bitbucket.org/aligent/aligent-graphql-mesh/pull-requests/446